Use internal timestamp for manual control setpoint.
The problem only occurs when using px4_ros_com when simulating the
flight of a vehicle on a PC.  The px4_ros_com client is very simple, so
the timestamp that is used to simulate RC communications using the
manual_control_setpoint message can only be supplied from the PC running
the simulation and the value of the PC timestamp can be much bigger than
the internal PX4 timestamp.

When the code in Commander.cpp gets the manual_control_setpoint message,
it compares the PC timestamp value against the internal timestamp using
the function hrt_elapsed_time().  When the larger PC timestamp is
subtracted from the smaller internal timestamp, the result is a massive
positive value due to the uint64_t wrapping around.  This causes the
code in Commander.cpp to think that the RC signal has been lost causing
an RTL fail safe.  This is not desirable when simulating the flight of a
vehicle!

This problem does not occur when using Mavlink as the Mavlink client
uses the internal timestamp when publishing the uORB message.